[INDDEX] Refactor code
